# Improve keys, keys_unsorted performance